Revert "Introduce range-loops"
This reverts commit 1f03224923cd1ef720da9454003ed6215a304a84.

Reason: the design is flawed. For example it doesn't allow natural
descending loops such as `for (a : 5, -5, -1) {}`
Reverting current implementation until it becomes widely used in
scripts to ease adoption of future, correctly designed version.
Proper design should behave similarly to iteration on ranges in Python.

Currently writing code requires knowledge of compiler internals and
this feature in result increases complexity instead of reducing it.