Allow to disable `switch` statement
`switch` statememnt obscures code flow - it needs too much attention
to notice (missing) breaks and to be aware of non-obvious side-effects
in the `case` branches code.
And performance-wise it is the same as `if`/`else` tree (it compiles
into the same bytecode).

When `switch` statement is disabled the keywords `switch`, `case` and
`default` may be used as identifiers.

This change makes error messages when using disabled keywords
less clear (should be understandable from the context though), but
allows to clean up the language syntax.