Drop support for := operator
It was incorrectly implemented during switching to AST-based compiler
But what's more important the initial idea was wrong - we don't
want a special operator that allows assignment inside expressions.
What we actually want is declaring and initializing local variables
inside `if` conditional expression just like in `for`.
This is to be implementated in future.