Compiler: don't mutate AST Id nodes from codegen
AST should be built by parser, probably it may be modified by
special transformation operations (such as optimizations).
But having nodes that change their properties ('isLocal'/'isAssignable')
depending on whether a codegen has been run on this AST or not is
conceptually wrong.

An identifier cannot be assignable or local or outer - a variable can.

We can manage and resolve variables by id in codegen itself, but
we already have a mechanism in SQFuncState (which is a codegen-stage
object), so use it for this.