Rework direct API for object comparison
Add sq_direct_is_equal() - this is not the same as invoking ObjCmp()
on objects. Logic of this comparison is the same as for <=> operator
and it can fail in certain scenarios.
Also for safety such comparison of non-numeric types (or classes
implementing _cmp metamethod) will be disabled soon.
In practice sq_direct_cmp() function was never needed - in all the
cases it was used for the equality check.
For this, implement sq_direct_is_equal() function.

Also simplify sq_direct_cmp() implementation - make it a thin wrapper
around SQVM::ObjCmp(). No need for extra logic that differs from VM
internals.