Detect incorrect arguments for imported functions
Argument / parameter checks didn't work for imported functions because
they are represented as SQObjects in external bindings instead of AST nodes.

Added external bindings as ExternalValueExpr in the static analyzer,
and made argument checks work with the SQObjects they hold.

Only functions and script closures are supported, native closures are not.