Fix the mess with debug line infos
_OP_LINE opcode was removed recently and this makes the naming
`_is_line_op` incorrect.
However even before that the naming was not the best, it indicated
the origin, but not the purpose. The functional meaning was whether
it should call a debugger callback on this line.
So, the better name is `_is_dbg_step_point`.

Also, after removing _OP_LINE there is no sense in not enabling line
information - this only sets the corresponding flag to false, but
provides no other benefits, the amount and complexity of the generated
code is the same.
Therefore this commits removes sq_enabledebuginfo() API function
(also a misleading name!), and lineInfo/debugInfo flags in various
places. No need to drag them around.

Now the naming is correct and meaningful and there are no useless
variables and functions around.